requirements.txt
asgiref==3.5.2
Django==3.2.12
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
